+++ backend/core/utils.py
@@ -12,6 +12,7 @@ import logging
 import requests
 import datetime
 import itertools
+from datetime import date
 
 from dateutil.relativedelta import relativedelta
 
@@ -21,9 +22,11 @@ from django.core.mail import EmailMessage
 from django.conf import settings
 from django.template.loader import render_to_string
 from django.core.mail import send_mail
+from django.utils import timezone
 from django.utils.translation import gettext_lazy as _
 from django.utils.html import strip_tags
 from django.db import transaction
+from django.db.models import Q
 
 from constance import config
 
@@ -419,7 +422,7 @@ def format_prism_payment_record(payment, line_no, record_no):
     16 - posteringshenvisningsnummer. As 117 in the finance records -
     date + machine number + record number - 8 chars + 5 chars + 7 chars.
 
-    17 - Payment ID. 20 digits with leading zeroes.
+    17 - The unique Payment pk. 20 digits with leading zeroes.
 
     40 - posting text.
     """
@@ -433,7 +436,7 @@ def format_prism_payment_record(payment, line_no, record_no):
         "10": "02",
         "11": f"{payment.recipient_id}",
         "12": f"{payment.date.strftime('%Y%m%d')}",
-        "17": f"{payment_id:020d}",
+        "17": f"{payment.pk:020d}",
         "40": f"Fra Ballerup Kommune ref: {payment_id}",
     }
     fields[
@@ -480,14 +483,14 @@ def export_prism_payments_for_date(date=None):
     """Process payments and output a file for PRISME."""
     # The output directory is not configurable - this is mapped through Docker.
     output_dir = settings.PRISM_OUTPUT_DIR
-    # Date = today if not given. We need "today" to set payment date.
-    today = datetime.datetime.now()
+    # Date = tomorrow if not given. We need "tomorrow" to set payment date.
+    tomorrow = datetime.datetime.now() + relativedelta(days=1)
     if not date:
-        date = today
+        date = tomorrow
 
-    # The microseconds are included to avoid accidentally overwriting today's
-    # file.
-    filename = f"{output_dir}/{date.strftime('%Y%m%d')}_{today.microsecond}"
+    # The microseconds are included to avoid accidentally overwriting
+    # tomorrow's file.
+    filename = f"{output_dir}/{date.strftime('%Y%m%d')}_{tomorrow.microsecond}"
     payments = due_payments_for_prism(date)
     if not payments.exists():
         # No payments
@@ -507,7 +510,7 @@ def export_prism_payments_for_date(date=None):
         trans_code = "Z300"  # Identifies transaction start.
 
         user_number = f"{config.PRISM_ORG_UNIT:04d}"  # Org unit.
-        day_of_year = today.timetuple().tm_yday  # Day of year.
+        day_of_year = tomorrow.timetuple().tm_yday  # Day of year.
 
         preamble_string = (
             f"{trans_code}{hdisp}{user_number}{media_type}{evolbr}"
@@ -526,7 +529,128 @@ def export_prism_payments_for_date(date=None):
     for p in payments:
         p.paid = True
         p.paid_amount = p.amount
-        p.paid_date = today
+        p.paid_date = tomorrow
         p.save()
     # Return filename for info and verification.
     return filename
+
+
+def generate_granted_payments_report_list():
+    """Generate a payments report of only granted payments."""
+    current_year = timezone.now().year
+    end_of_current_year = date.max.replace(year=current_year)
+    two_years_ago = current_year - 2
+    beginning_of_two_years_ago = date.min.replace(year=two_years_ago)
+
+    granted_activities = models.Activity.objects.filter(
+        status=models.STATUS_GRANTED
+    )
+    payment_ids = granted_activities.values_list(
+        "payment_plan__payments__pk", flat=True
+    )
+    payments = (
+        models.Payment.objects.filter(id__in=payment_ids)
+        .paid_date_or_date_gte(beginning_of_two_years_ago)
+        .paid_date_or_date_lte(end_of_current_year)
+        .select_related(
+            "payment_schedule__activity__appropriation__case",
+            "payment_schedule__activity__appropriation__section",
+            "payment_schedule__activity__details",
+        )
+    )
+    payments_report_list = generate_payments_report_list(payments)
+    return payments_report_list
+
+
+def generate_expected_payments_report_list():
+    """Generate a payments report of granted AND expected payments."""
+    current_year = timezone.now().year
+    end_of_current_year = date.max.replace(year=current_year)
+    two_years_ago = current_year - 2
+    beginning_of_two_years_ago = date.min.replace(year=two_years_ago)
+
+    expected_activities = models.Activity.objects.filter(
+        Q(status=models.STATUS_GRANTED) | Q(status=models.STATUS_EXPECTED)
+    )
+    payment_ids = [
+        payment.id
+        for activity in expected_activities
+        for payment in activity.applicable_payments
+    ]
+    payments = (
+        models.Payment.objects.filter(id__in=payment_ids)
+        .paid_date_or_date_gte(beginning_of_two_years_ago)
+        .paid_date_or_date_lte(end_of_current_year)
+        .select_related(
+            "payment_schedule__activity__appropriation__case",
+            "payment_schedule__activity__appropriation__section",
+            "payment_schedule__activity__details",
+        )
+    )
+    payments_report_list = generate_payments_report_list(payments)
+    return payments_report_list
+
+
+def generate_payments_report_list(payments):
+    """Generate a payments report list of payment dicts from payments."""
+    payments_report_list = []
+    for payment in payments:
+        try:
+            activity = payment.payment_schedule.activity
+        except models.Activity.DoesNotExist:
+            logger.exception(
+                f"PaymentSchedule {payment.payment_schedule.pk}"
+                f" has no activity"
+            )
+            continue
+
+        case = activity.appropriation.case
+        appropriation = activity.appropriation
+        payment_schedule = payment.payment_schedule
+
+        main_activity_id = (
+            appropriation.main_activity.details.activity_id
+            if appropriation.main_activity
+            else None
+        )
+
+        payment_dict = {
+            # payment specific.
+            "id": payment.pk,
+            "amount": payment.amount,
+            "paid_amount": payment.paid_amount,
+            "date": payment.date,
+            "paid_date": payment.paid_date,
+            "account_string": payment.account_string,
+            # payment_schedule specific.
+            "payment_schedule__"
+            "payment_amount": payment_schedule.payment_amount,
+            "payment_schedule__"
+            "payment_frequency": payment_schedule.payment_frequency,
+            "recipient_type": payment_schedule.recipient_type,
+            "recipient_id": payment_schedule.recipient_id,
+            "recipient_name": payment_schedule.recipient_name,
+            "payment_method": payment_schedule.payment_method,
+            # activity specific.
+            "activity__details__activity_id": activity.details.activity_id,
+            "activity__details__name": activity.details.name,
+            "activity_start_date": activity.start_date,
+            "activity_end_date": activity.end_date,
+            # appropriation specific.
+            "section": str(appropriation.section),
+            "sbsys_id": appropriation.sbsys_id,
+            "main_activity_id": main_activity_id,
+            # case specific.
+            "cpr_number": case.cpr_number,
+            "name": case.name,
+            "target_group": case.target_group,
+            "case_worker": str(case.case_worker),
+            "team": str(case.team) if case.team else None,
+            "leader": str(case.team.leader) if case.team else None,
+            "effort_step": str(case.effort_step),
+            "scaling_step": str(case.scaling_step),
+            "paying_municipality": str(case.paying_municipality),
+        }
+        payments_report_list.append(payment_dict)
+
+    return payments_report_list