+++ NEWS.rst
@@ -1,19 +1,45 @@
-* Version 2.3.0, 2020-01-09
+Version 2.4.0, 2020-01-24
+-------------------------
 
 New in this version:
 
-Features:
+Features
+^^^^^^^^
+
+* Add warning in GUI if a payment date is earlier than two days from today.
+* Add restriction in GUI so an Activity can only have one expected Activity.
+* Add Actual-state CSV generation for Payments.
+* Improve documentation all-around.
+* Change payment file default date to tomorrow.
+* Update Django from 2.2.4 to 2.2.9
+
+Bug fixes
+^^^^^^^^^
+
+* Fix CPR search for "Find sager".
+* Change field 17 of PRISM file to the unique Payment pk.
+* Handle missing effort steps gracefully in GUI.
+
+Version 2.3.0, 2020-01-09
+-------------------------
+
+New in this version:
+
+Features
+^^^^^^^^
 
 * Modify the URL for the rate tabel (taksttabel) to a more general one.
 * Nice-ify django admin for Payments and PaymentSchedules and allow search on payment id.
 * Add pydocstyle compliance.
-* Add sphinx docs generation
+* Add sphinx docs generation.
 
-Bug fixes:
+Bug fixes
+^^^^^^^^^
 
 * Fix incorrect tests dependent on current year.
 
-* Version 2.2.3, 2019-12-12
+Version 2.2.3, 2019-12-12
+-------------------------
 
 New in this version:
 
@@ -32,22 +58,27 @@ New in this version:
 * Fix paths in settings.
 * Update Django from 2.2.1 to 2.2.4
 
-* Version 2.2.2, 2019-11-28
+Version 2.2.2, 2019-11-28
+-------------------------
 
-Hotfix:
+Hotfix
+^^^^^^
 
 * Fix broken migration.
 
 
-* Version 2.2.1, 2019-11-25
+Version 2.2.1, 2019-11-25
+-------------------------
 
 New in this version:
 
-Features:
+Features
+^^^^^^^^
 
 * Mark payments for SD Løn along with fictive ones.
 
-Bug fixes:
+Bug fixes
+^^^^^^^^^
 
 * Fix hover text.
 * Display of Indsatstrappen fixed.
@@ -57,11 +88,13 @@ Bug fixes:
 * Empty "not found" text when displaying "Mine sager".
 
 
-* Version 2.2.0, 2019-11-21
+Version 2.2.0, 2019-11-21
+-------------------------
 
 New in this version:
 
-Features:
+Features
+^^^^^^^^
 
 * It is now possible to find payments from a payment ID.
 * Case worker can now be changed on several cases in one action.
@@ -92,7 +125,8 @@ Features:
 * Information about citizen included in display of appropriation.
 
 
-Bug fixes:
+Bug fixes
+^^^^^^^^^
 
 * Don't throw an exception if users attempt to access the API without
   logging in, just deny access.
@@ -104,14 +138,16 @@ Bug fixes:
   activities.
 
 
-* Version 2.0.1, 2019-11-11
+Version 2.0.1, 2019-11-11
+-------------------------
 
 New in this version:
 
 * Add support for Service Provider certificates through PySaml2.
 
 
-* Version 2.0.0, 2019-11-06
+Version 2.0.0, 2019-11-06
+-------------------------
 
 New in this version:
 
@@ -131,7 +167,8 @@ New in this version:
 * Fix redirect when setting a payment paid.
 * Add missing verbose names in Django admin.
 
-* Version 1.1.1, 2019-10-30
+Version 1.1.1, 2019-10-30
+-------------------------
 
 Hotfix release.
 
@@ -144,7 +181,8 @@ New in this version:
 * Fix invalid XML in OS2forms.xml.
 * Add missing constraint for creating supplementary activities based on allowed main activities.
 
-* Version 1.1.0, 2019-10-04
+Version 1.1.0, 2019-10-04
+-------------------------
 
 New in this version:
 
@@ -163,7 +201,8 @@ New in this version:
 * Suppress not writeable warning from ipython.
 
 
-* Version 1.0.0, 2019-09-27
+Version 1.0.0, 2019-09-27
+-------------------------
 
 First production release. New in this version:
 
@@ -200,7 +239,8 @@ First production release. New in this version:
 * DB representation of effort steps (Indsatstrappe) changed to integer.
 
 
-* Version 0.5.0, 2019-09-05
+Version 0.5.0, 2019-09-05
+-------------------------
 
 New in this version: